# k8s/src/appstore-api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AppStore API",
    "description": "App catalogue, deployments and teams for the DHBW cloud self-service platform",
    "version": "0.1.0-test.1"
  },
  "paths": {
...
    "/apps/{app_id}": {
      "get": {
        "tags": [
          "Apps"
        ],
        "summary": "Get App",
        "description": "Get one app with the versions (git tags) the caller may deploy.\n\nThe owner and admins see every tag, each flagged ``approved`` when an\nadmin approved the commit the tag points to now; everyone else sees\nonly approved tags. If the repository cannot be read, ``versions`` is\nempty rather than the request failing.\n\nSoft-deleted apps are still readable here so existing deployments\nthat still reference them can render their app name, git link,\netc.; their ``versions`` list is always empty. They just don't show\nup in the apps list / deploy wizard.\n\n- 403: the app is private or has no approved version, and the caller\n  is neither owner nor admin.\n- 404: no such app.",
        "operationId": "getApp",
        "parameters": [
          {
            "name": "app_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "App Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppWithVersions"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Apps"
        ],
        "summary": "Update App",
        "description": "Update an app's name, description, visibility (``is_private``) or image.\n\nOnly fields sent in the body change; ``image: \"\"`` clears the image.\n``git_link`` is immutable after creation and ignored if sent.\n\nOwner or admin only (403 otherwise). An app an admin deactivated\n(``hidden_by_admin``) can be made public again only by an admin (403\n``app_hidden_by_admin`` for the owner); an admin doing so lifts the\ndeactivation. 404 if the app does not exist or is deleted.",
        "operationId": "updateApp",
        "parameters": [
          {
            "name": "app_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "App Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Apps"
        ],
        "summary": "Delete App",
        "description": "Soft-delete an app.\n\nSets ``apps.deleted_at`` instead of removing the row, so any\nhistorical or still-running deployment that points at this app keeps\nresolving. The app stops appearing in listings and the deploy wizard;\nexisting deployments live on until destroyed individually.\n\nOwner or admin only (403 otherwise). 404 if the app does not exist or\nis already deleted.",
        "operationId": "deleteApp",
        "parameters": [
          {
            "name": "app_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "App Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/apps/{app_id}/versions/{version_tag}/approve": {
      "post": {
        "tags": [
          "Admin",
          "Admin"
        ],
        "summary": "Approve Version",
        "description": "Approve a submitted version, making it deployable by all users.\n\nAdmins only (403 otherwise). Works on PENDING and on previously REJECTED\nsubmissions.\n\n- 404: the app or the version submission does not exist.\n- 409 ``version_moved``: the tag no longer points to the submitted commit;\n  the admin would be approving code that was not what got submitted.\n- 409: the version is already approved.\n- 422: the ``@openstack`` markers in this version's Terraform/Packer\n  variable files are invalid (``marker_errors`` lists them). Checked with\n  the same helper as ``GET /apps/{id}/variables``, on the submitted commit.\n- 502: the repository cannot be read; nothing is approved unchecked.",
        "operationId": "approveVersion",
        "parameters": [
          {
            "name": "app_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "App Id"
            }
          },
          {
            "name": "version_tag",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Version Tag"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppVersionApprovalResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/apps/{app_id}/versions/{version_tag}/reject": {
      "post": {
        "tags": [
          "Admin",
          "Admin"
        ],
        "summary": "Reject Version",
        "description": "Reject a PENDING version with a mandatory reason shown to the app owner.\n\nAdmins only (403 otherwise). 404 if the app or submission does not exist,\n409 if the version is not PENDING (use ``/revoke`` for approved ones).",
        "operationId": "rejectVersion",
        "parameters": [
          {
            "name": "app_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "App Id"
            }
          },
          {
            "name": "version_tag",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Version Tag"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppVersionApprovalDecision"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppVersionApprovalResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/apps/{app_id}/versions/{version_tag}/revoke": {
      "post": {
        "tags": [
          "Admin",
          "Admin"
        ],
        "summary": "Revoke Version",
        "description": "Revoke a previously APPROVED version with a mandatory reason (sets status to REJECTED).\n\nA public app can no longer be deployed with this version. Admins only (403 otherwise). 404 if the app or submission\ndoes not exist, 409 if the version is not APPROVED.",
        "operationId": "revokeVersion",
        "parameters": [
          {
            "name": "app_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "App Id"
            }
          },
          {
            "name": "version_tag",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Version Tag"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppVersionApprovalDecision"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppVersionApprovalResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/apps/{app_id}": {
      "put": {
        "tags": [
          "Admin",
          "Admin"
        ],
        "summary": "Deactivate App",
        "description": "Emergency deactivation: set the app to private so it disappears from the store.\n\nTakes no body. Does not delete the app, its versions or its deployments.\nMarks it ``hidden_by_admin``: the owner can no longer make it public,\nonly an admin can (``PUT /apps/{id}`` with ``is_private: false``).\nAdmins only (403 otherwise), 404 if the app does not exist (soft-deleted\napps included).",
        "operationId": "deactivateApp",
        "parameters": [
          {
            "name": "app_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "App Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AppResponse": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "git_link": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Git Link"
          },
          "is_private": {
            "type": "boolean",
            "title": "Is Private"
          },
          "appId": {
            "type": "string",
            "format": "uuid",
            "title": "Appid"
          },
          "userId": {
            "type": "string",
            "format": "uuid",
            "title": "Userid"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "hidden_by_admin": {
            "type": "boolean",
            "title": "Hidden By Admin",
            "default": false
          },
          "image": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image"
          }
        },
        "type": "object",
        "required": [
          "name",
          "is_private",
          "appId",
          "userId",
          "created_at"
        ],
        "title": "AppResponse",
        "description": "An app of the catalogue."
      },
      "AppUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "is_private": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Private"
          },
          "image": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image"
          }
        },
        "type": "object",
        "title": "AppUpdate",
        "description": "Body of ``PUT /apps/{id}``: partial update; omitted fields stay unchanged."
      },
...
      "AppVersionApprovalDecision": {
        "properties": {
          "rejection_reason": {
            "type": "string",
            "title": "Rejection Reason"
          }
        },
        "type": "object",
        "required": [
          "rejection_reason"
        ],
        "title": "AppVersionApprovalDecision",
        "description": "Body of reject and revoke: the reason, shown to the app owner."
      },
      "AppVersionApprovalResponse": {
        "properties": {
          "approvalId": {
            "type": "string",
            "format": "uuid",
            "title": "Approvalid"
          },
          "appId": {
            "type": "string",
            "format": "uuid",
            "title": "Appid"
          },
          "version_tag": {
            "type": "string",
            "title": "Version Tag"
          },
          "commit_sha": {
            "type": "string",
            "title": "Commit Sha"
          },
          "status": {
            "$ref": "#/components/schemas/AppVersionApprovalStatus"
          },
          "diff_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Diff Url"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "rejection_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rejection Reason"
          },
          "reviewed_by": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewed By"
          },
          "reviewed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewed At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "approvalId",
          "appId",
          "version_tag",
          "commit_sha",
          "status",
          "created_at"
        ],
        "title": "AppVersionApprovalResponse",
        "description": "The review state of one version tag of an app, bound to the reviewed commit."
      },
      "AppVersionApprovalStatus": {
        "type": "string",
        "enum": [
          "pending",
          "approved",
          "rejected"
        ],
        "title": "AppVersionApprovalStatus",
        "description": "Review state of an app version; a public app deploys only APPROVED versions (plan E6)."
      },
...
      "AppWithVersions": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "git_link": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Git Link"
          },
          "is_private": {
            "type": "boolean",
            "title": "Is Private"
          },
          "appId": {
            "type": "string",
            "format": "uuid",
            "title": "Appid"
          },
          "userId": {
            "type": "string",
            "format": "uuid",
            "title": "Userid"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "hidden_by_admin": {
            "type": "boolean",
            "title": "Hidden By Admin",
            "default": false
          },
          "image": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image"
          },
          "user": {
            "$ref": "#/components/schemas/UserResponse"
          },
          "versions": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Versions",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "name",
          "is_private",
          "appId",
          "userId",
          "created_at",
          "user"
        ],
        "title": "AppWithVersions",
        "description": "An app with its owner and its Git tags, each marked with its approval state."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "UserResponse": {
        "properties": {
          "userId": {
            "type": "string",
            "format": "uuid",
            "title": "Userid"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "firstName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Firstname"
          },
          "lastName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lastname"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "userId",
          "email",
          "username",
          "created_at"
        ],
        "title": "UserResponse",
        "description": "A person as other resources reference them. Rights are not part of it:\nthey come from the caller's tokens and are never stored."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}